pyproject.toml
[project]
name = "goofish-watcher"
version = "0.1.0"
description = "Discord bot that monitors Xianyu/Goofish listings with AI verification"
requires-python = ">=3.11"

dependencies = [
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "playwright>=1.40.0",
    "aiofiles>=23.2.1",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
include = ["bot*", "core*", "db*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
